+++ src/test/java/io/github/astrapi69/mystic/crypt/key/SharedSecretEncryptorTest.java
@@ -10,6 +10,7 @@ import java.security.SecureRandom;
 import java.security.Security;
 
 import io.github.astrapi69.crypt.data.factory.KeyPairGeneratorFactory;
+import io.github.astrapi69.crypt.data.model.SharedSecretModel;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.testng.annotations.Test;
 
@@ -17,7 +18,6 @@ import io.github.astrapi69.crypt.data.factory.KeyPairFactory;
 
 public class SharedSecretEncryptorTest
 {
-
 
 	/**
 	 * Test method for {@link PublicKeyEncryptor} constructor with {@link PublicKey} object
@@ -33,8 +33,12 @@ public class SharedSecretEncryptorTest
 		SharedSecretEncryptor encryptor;
 		SharedSecretDecryptor decryptor;
 		byte[] encrypted;
+		byte[] decrypt;
 		String longString;
 		byte[] iv;
+		SharedSecretModel aliceSharedSecretModel;
+		SharedSecretModel bobSharedSecretModel;
+
 		Security.addProvider(new BouncyCastleProvider());
 
 		iv = new SecureRandom().generateSeed(16);
@@ -43,7 +47,6 @@ public class SharedSecretEncryptorTest
 			.newKeyPairGenerator("brainpoolp256r1", "ECDH", "BC").generateKeyPair();
 		KeyPair keyPairAlice = KeyPairGeneratorFactory
 			.newKeyPairGenerator("brainpoolp256r1", "ECDH", "BC").generateKeyPair();
-
 
 		encryptor = new SharedSecretEncryptor(keyPairBob.getPrivate(), keyPairAlice.getPublic(),
 			"ECDH", "AES", "BC", "AES/GCM/NoPadding", iv);
@@ -56,7 +59,42 @@ public class SharedSecretEncryptorTest
 		decryptor = new SharedSecretDecryptor(keyPairAlice.getPrivate(), keyPairBob.getPublic(),
 			"ECDH", "AES", "BC", "AES/GCM/NoPadding", iv);
 
-		byte[] decrypt = decryptor.decrypt(encrypted);
+		decrypt = decryptor.decrypt(encrypted);
+		actual = new String(decrypt, "UTF-8");
+
+		assertEquals(expected, actual);
+
+		// with model
+
+		bobSharedSecretModel = SharedSecretModel.builder()
+			.privateKey(keyPairBob.getPrivate())
+			.publicKey(keyPairAlice.getPublic())
+			.keyAgreementAlgorithm("ECDH")
+			.secretKeyAlgorithm("AES")
+			.provider("BC")
+			.cipherTransformation("AES/GCM/NoPadding")
+			.iv(iv).build();
+
+
+		encryptor = new SharedSecretEncryptor(bobSharedSecretModel);
+
+		expected = "Hi there, whats up!";
+
+		encrypted = encryptor.encrypt(expected.getBytes(StandardCharsets.UTF_8));
+		assertNotNull(encrypted);
+
+		aliceSharedSecretModel = SharedSecretModel.builder()
+			.privateKey(keyPairAlice.getPrivate())
+			.publicKey(keyPairBob.getPublic())
+			.keyAgreementAlgorithm("ECDH")
+			.secretKeyAlgorithm("AES")
+			.provider("BC")
+			.cipherTransformation("AES/GCM/NoPadding")
+			.iv(iv).build();
+
+		decryptor = new SharedSecretDecryptor(aliceSharedSecretModel);
+
+		decrypt = decryptor.decrypt(encrypted);
 		actual = new String(decrypt, "UTF-8");
 
 		assertEquals(expected, actual);